refactor: extract methods and components from EditTag
